+++ lib/absinthe/phase/schema/validation/object_interfaces_must_be_valid.ex
@@ -13,8 +13,7 @@ defmodule Absinthe.Phase.Schema.Validation.ObjectInterfacesMustBeValid do
     ifaces =
       schema.type_definitions
       |> Enum.filter(&match?(%Blueprint.Schema.InterfaceTypeDefinition{}, &1))
-      |> Enum.map(& &1.identifier)
-      |> MapSet.new()
+      |> Map.new(&{&1.identifier, &1})
 
     schema = Blueprint.prewalk(schema, &validate_objects(&1, ifaces))
     {:halt, schema}
@@ -24,23 +23,57 @@ defmodule Absinthe.Phase.Schema.Validation.ObjectInterfacesMustBeValid do
     obj
   end
 
-  defp validate_objects(%Blueprint.Schema.ObjectTypeDefinition{} = object, ifaces) do
-    Enum.reduce(object.interfaces, object, fn iface, object ->
-      if iface in ifaces do
-        object
-      else
-        detail = %{
-          object: object.identifier,
-          interface: iface
-        }
-
-        object |> put_error(error(object, detail))
-      end
-    end)
+  defp validate_objects(%struct{} = object, all_interfaces)
+       when struct in [
+              Blueprint.Schema.ObjectTypeDefinition,
+              Blueprint.Schema.InterfaceTypeDefinition
+            ] do
+    check_transitive_interfaces(object, object.interfaces, all_interfaces, nil, [])
   end
 
   defp validate_objects(type, _) do
     type
+  end
+
+  # check that the object declares it implements all interfaces up the
+  # hierarchy chain as per spec https://github.com/graphql/graphql-spec/blame/October2021/spec/Section%203%20--%20Type%20System.md#L1158-L1161
+  defp check_transitive_interfaces(
+         object,
+         [object_interface | tail],
+         all_interfaces,
+         implemented_by,
+         visited
+       ) do
+    current_interface = all_interfaces[object_interface]
+
+    if current_interface && current_interface.identifier in object.interfaces do
+      case current_interface do
+        %{interfaces: interfaces} = interface ->
+          # to prevent walking in cycles we need to filter out visited interfaces
+          interfaces = Enum.filter(interfaces, &(&1 not in visited))
+
+          check_transitive_interfaces(object, tail ++ interfaces, all_interfaces, interface, [
+            object_interface | visited
+          ])
+
+        _ ->
+          check_transitive_interfaces(object, tail, all_interfaces, implemented_by, [
+            object_interface | visited
+          ])
+      end
+    else
+      detail = %{
+        object: object.identifier,
+        interface: object_interface,
+        implemented_by: implemented_by
+      }
+
+      object |> put_error(error(object, detail))
+    end
+  end
+
+  defp check_transitive_interfaces(object, [], _, _, _) do
+    object
   end
 
   defp error(object, data) do
@@ -58,9 +91,19 @@ defmodule Absinthe.Phase.Schema.Validation.ObjectInterfacesMustBeValid do
   Reference: https://github.com/facebook/graphql/blob/master/spec/Section%203%20--%20Type%20System.md#interfaces
   """
 
-  def explanation(%{object: obj, interface: interface}) do
+  def explanation(%{object: obj, interface: interface, implemented_by: nil}) do
     """
-    Type "#{obj}" cannot implement non-interface type "#{interface}"
+    Type "#{obj}" must implement interface type "#{interface}"
+
+    #{@description}
+    """
+  end
+
+  def explanation(%{object: obj, interface: interface, implemented_by: implemented}) do
+    """
+    Type "#{obj}" must implement interface type "#{interface}" because it is implemented by "#{
+      implemented.identifier
+    }".
 
     #{@description}
     """