+++ test/absinthe/schema/rule/object_must_implement_interfaces_test.exs
@@ -7,6 +7,7 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
     object :user do
       interface :named
       interface :favorite_foods
+      interface :parented
       field :name, :string
       field :id, :id
       field :parent, :named
@@ -22,6 +23,13 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
     interface :parented do
       field :parent, :named
       field :another_parent, :named
+
+      resolve_type fn
+        %{type: :dog}, _ -> :dog
+        %{type: :user}, _ -> :user
+        %{type: :cat}, _ -> :cat
+        _, _ -> nil
+      end
     end
 
     interface :named do
@@ -52,6 +60,7 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
     object :dog do
       field :name, :string
       interface :named
+      interface :parented
       interface :favorite_foods
       field :parent, :named
       field :another_parent, :user
@@ -63,6 +72,7 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
     object :cat do
       interface :named
       interface :favorite_foods
+      interface :parented
       field :name, non_null(:string)
       field :parent, :named
       field :another_parent, :user
@@ -89,7 +99,7 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
     assert %{
              named: [:cat, :dog, :user],
              favorite_foods: [:cat, :dog, :user],
-             parented: [:named]
+             parented: [:cat, :dog, :user]
            } ==
              Schema.__absinthe_interface_implementors__()
   end
@@ -107,7 +117,7 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
       url: String
     }
 
-    interface Image implements Resource & Node {
+    type Image implements Resource & Node {
       id: ID!
       url: String
       thumbnail: String
@@ -115,14 +125,21 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfacesTest do
 
     """
 
+    def hydrate(%Absinthe.Blueprint.Schema.InterfaceTypeDefinition{}, _) do
+      {:resolve_type, &__MODULE__.resolve_type/1}
+    end
+
+    def hydrate(_node, _ancestors), do: []
+
+    def resolve_type(_), do: false
+
     query do
     end
   end
 
   test "interfaces are set from sdl" do
     assert %{
-             image: [],
-             node: [:image, :resource],
+             node: [:image],
              resource: [:image]
            } ==
              InterfaceImplementsInterfaces.__absinthe_interface_implementors__()