+++ test/support/fixtures/dynamic/invalid_output_types_sdl.exs
@@ -1,0 +1,17 @@
+defmodule Absinthe.Fixtures.InvalidInputTypesSdlSchema do
+  use Absinthe.Schema
+
+  import_sdl """
+    type User
+
+    input Input
+
+    type BadObject {
+      blah: Input
+    }
+
+    type Query {
+      foo(invalidArg: User): User
+    }
+  """
+end